fix: use carriage returns in response (#7)
* fix: use carriage returns in response

The HTTP 1.1 standard requires header fields (and blank lines) to be terminated in CRLF (\r\n).
Some http request libraries have taken this to heart and throw errors when this is not the case, or require a special
flag to bypass the issue. Reference: https://www.rfc-editor.org/rfc/rfc2616#section-2.2

Examples of libraries that enforce this in node.js: axios, http